requirements.txt
discord.py[voice]
python-dotenv
yt-dlp
spotipy
PyNaCl
wavelink>=3.0.0,<4.0.0